+++ src/features/admin/api/deleteClub.ts
